+++ src/pages/Product/Product.js
@@ -1,30 +1,55 @@
 import React from "react";
+import { useState } from "react";
 import "./Product.scss";
+import ColorButton from "./components/ColorButton/ColorButton";
+import Color from "./components/Color/Color";
+import Count from "./components/Count/Count";
+import Review from "./components/Review/Review";
 
 const Product = () => {
   const price = 300;
-  const totalPrice = price;
+  const [count, setCount] = useState(1);
+  const handleIncreaseBtn = () => {
+    setCount(count + 1);
+  };
+  const handleDecreaseBtn = () => {
+    if (count - 1) {
+      setCount(count - 1);
+    }
+  };
+  const handleResetBtn = () => {
+    setCount(1);
+  };
+  const [color, setColor] = useState("white");
+  const colorHandlerButton = (e) => {
+    setColor(e);
+  };
 
   return (
     <div className="product">
       <div className="productDetail">
         <div className="productDetailImg">
           <img
-            src={`/images/golf-ball-.jpg`} // color 이름에 따라 다른 이미지 경로 넣기
+            src={`/images/golf-ball-${color}.jpg`} // color 이름에 따라 다른 이미지 경로 넣기
             alt={`golf-ball`}
           />
-          {/* ColorButton 컴포넌트 위치 */}
+          <ColorButton color={color} />
         </div>
         <div className="productDetailInfo">
           <span className="title">골프공</span>
           <span>비거리를 비약적으로 늘려줍니다</span>
           <span>가격 : {price.toLocaleString()} 원</span>
-          {/* Color 컴포넌트 위치 */}
+          <Color color={color} colorHandlerButton={colorHandlerButton} />
           <div className="quantity">
             <span> 수량 : </span>
-            {/* Count 컴포넌트 위치 */}
+            <Count
+              count={count}
+              handleIncreaseBtn={handleIncreaseBtn}
+              handleDecreaseBtn={handleDecreaseBtn}
+              handleResetBtn={handleResetBtn}
+            />
           </div>
-          <span>최종 가격 : {totalPrice.toLocaleString()} 원</span>
+          <span>최종 가격 : {count * price} 원</span>
           <button className="buyBtn">구매하기</button>
         </div>
       </div>
@@ -32,7 +57,7 @@ const Product = () => {
         <div className="reviewListHeader">
           <span>상품평</span>
         </div>
-        {/* Review 컴포넌트 위치 */}
+        <Review />
       </div>
     </div>
   );
