+++ src/pages/Product/components/ColorButton/ColorButton.scss
@@ -1,0 +1,21 @@
+@mixin colors {
+  &.white {
+    background-color: white;
+  }
+
+  &.red {
+    background-color: #f32321;
+  }
+
+  &.yellow {
+    background-color: #faf1b0;
+  }
+}
+
+div .colorButton {
+  @include colors;
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  cursor: pointer;
+}
